Fix NetStandard issue in System.IO.FileSystem.AccessControl
