+++ src/Common/src/CoreLib/System/IO/PathInternal.cs
@@ -18,12 +18,9 @@ namespace System.IO
 #if MS_IO_REDIST
         internal static string EnsureTrailingSeparator(string path)
             => EndsInDirectorySeparator(path) ? path : path + DirectorySeparatorCharAsString;
-
-        internal static bool EndsInDirectorySeparator(string path)
-            => !string.IsNullOrEmpty(path) && IsDirectorySeparator(path[path.Length - 1]);
 #else
         internal static string EnsureTrailingSeparator(string path)
-            => Path.EndsInDirectorySeparator(path.AsSpan()) ? path : path + DirectorySeparatorCharAsString;
+            => EndsInDirectorySeparator(path.AsSpan()) ? path : path + DirectorySeparatorCharAsString;
 #endif
 
         internal static bool IsRoot(ReadOnlySpan<char> path)
@@ -219,5 +216,33 @@ namespace System.IO
 
             return true;
         }
+
+        /// <summary>
+        /// Trims one trailing directory separator beyond the root of the path.
+        /// </summary>
+        internal static string TrimEndingDirectorySeparator(string path) =>
+            EndsInDirectorySeparator(path) && !IsRoot(path.AsSpan()) ?
+                path.Substring(0, path.Length - 1) :
+                path;
+
+        /// <summary>
+        /// Returns true if the path ends in a directory separator.
+        /// </summary>
+        internal static bool EndsInDirectorySeparator(string path)
+              => !string.IsNullOrEmpty(path) && IsDirectorySeparator(path[path.Length - 1]);
+
+        /// <summary>
+        /// Trims one trailing directory separator beyond the root of the path.
+        /// </summary>
+        internal static ReadOnlySpan<char> TrimEndingDirectorySeparator(ReadOnlySpan<char> path) =>
+            EndsInDirectorySeparator(path) && !IsRoot(path) ?
+                path.Slice(0, path.Length - 1) :
+                path;
+
+        /// <summary>
+        /// Returns true if the path ends in a directory separator.
+        /// </summary>
+        internal static bool EndsInDirectorySeparator(ReadOnlySpan<char> path)
+            => path.Length > 0 && IsDirectorySeparator(path[path.Length - 1]);
     }
 }
