+++ src/Common/src/System/IO/FileSystem.Attributes.Windows.cs
@@ -57,7 +57,7 @@ namespace System.IO
             int errorCode = Interop.Errors.ERROR_SUCCESS;
 
             // Neither GetFileAttributes or FindFirstFile like trailing separators
-            path = Path.TrimEndingDirectorySeparator(path);
+            path = PathInternal.TrimEndingDirectorySeparator(path);
 
             using (DisableMediaInsertionPrompt.Create())
             {
