+++ src/Common/src/System/IO/FileSystem.DirectoryCreation.Windows.cs
@@ -42,7 +42,7 @@ namespace System.IO
             int length = fullPath.Length;
 
             // We need to trim the trailing slash or the code will try to create 2 directories of the same name.
-            if (length >= 2 && Path.EndsInDirectorySeparator(fullPath.AsSpan()))
+            if (length >= 2 && PathInternal.EndsInDirectorySeparator(fullPath.AsSpan()))
             {
                 length--;
             }
